+++ Validator/UniqueValidator.php
@@ -40,8 +40,8 @@ class UniqueValidator extends ConstraintValidator
     /**
      * Checks if the passed value is valid.
      *
-     * @param mixed      $value      The value that should be validated
-     * @param Constraint $constraint The constrain for the validation
+     * @param UserInterface $value      The value that should be validated
+     * @param Constraint    $constraint The constrain for the validation
      *
      * @return Boolean Whether or not the value is valid
      *