+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+### v1.5.2  (2012-02-18)
+
+* Add support for Composer package manager now you can find this bundle in http://www.packagist.org
+* [Security] Sanitize user input.
+* Other fixes in sync with Symfony updates.
+
+### v1.5.1  (2012-02-06)
+
+* [Security] Prevent accidental information disclosure
+
 ### v1.5.0  (2012-01-29)
 
 * Enhancement release